fix: resolve KeychainHelper scope issue causing Swift compilation failure
Issue: SwiftCompile failed with "Cannot find 'KeychainHelper' in scope"
Root Cause: Separate KeychainHelper.swift file not being found by compiler

Solution:
- Move Keychain operations inline to AuthState class as private methods
- Add storeJWTToken(), getJWTToken(), and deleteJWTToken() methods directly to AuthState
- Remove dependency on separate KeychainHelper.swift file
- Update all keychainHelper.method() calls to use inline methods
- Maintain same Keychain functionality with proper Security framework usage

Benefits:
- Eliminates compilation scope issues
- Reduces file dependencies
- Keeps Keychain operations encapsulated within AuthState
- Same secure token storage functionality
- No external class dependencies
